# WIP: optimising MakeAppointments runtime 