I have used dictionary returns for `grant_burn` to calculate the burn of grants only once at the start of the program and access grant amounts in O(1) time

This has reduced runtime for MakeAppointments on the (small) grants dbs in the schema exemplars from 17s 616ms to 12s 524 ms, by a little under 40%